types: split check.go (3705 lines) into 5 files
Pure code motion. Splits into:
- types/kinds.go (245 lines): Type interface, kind structs, String methods, Subst.
- types/unify.go (257 lines): the unify algorithm.
- types/resolve.go (125 lines): resolveTypeRef + helpers.
- types/check.go (1121 lines): Check entry + checkStmt + checkIfStmt.
- types/check_expr.go (2000 lines): all expression checkers + alias helpers.

No behaviour change; full suite passes. Closes CLEAN-5 and CLEAN-13.
